+++ yiyingliService/src/cn/yiyingli/Dao/TeacherDao.java
@@ -22,7 +22,7 @@ public interface TeacherDao {
 
 	void update(Teacher teacher);
 
-	void updateAddMile(long teacherId, long mile);
+	void updateAddMile(long teacherId, float mile);
 
 	void updateAddSubMile(long teacherId, long subMile);
